Refactor time stuff into Utils, accept all timezones
